Added legacy YouTube username support
Added ability to retrieve playlists by suppyling a legacy YouTube
username via the -y option. This entails first performing a channel
request using the forUsername parameter, then performing a normal
playlists request using the id part from the channel response.

A try-except clause is used in the case that the username is invalid. An
alternate method of handling this case would be to check the
pageInfo.totalResults property in the response for zero.