Moved backup_playlists() call to main()
Instead of calling backup_playslists() from within setup_request(), it
is now called from main(). setup_request() returns the request, and
backup_playlists() accepts the request as an argument. This further
distinguishes the roles of both functions and results in a clearer
function call order.

In order to achieve this, it was necessary to make the youtube API
resource object global so that it can be accessed from both functions.