Changed output from files to stdout
In order to make ypb a filter (read input from stdin and produce outputs
on stdout) writing output to files was removed in favor of printing to
stdout. This is method better supports shell integration features such
as IO redirection and pipelining. As a result, code complexity is
reduced but text formatters such as awk, sed, or csplit are required
should the user wish to modify the output.