Added credentials refresh for authorized requests
Added check to see if access token is expired, and if so, refresh the
credentials using the refresh token.

This is especially important for automated circumstances that do not
have the capability of retrieving credentials interactively through a
web browser.

Prior implementation merely re-ran the interactive authorization
process upon token expiration, thus hanging for scheduled scripts.